Deprecate nvCOMP python API (#714)
Closes #685 

This PR marks nvCOMP python bindings deprecated.

Specifically, this PR adds a free function `kvikio_deprecate_module` used to deprecate modules. This PR also improves the decorator `@kvikio_deprecation_notice`, which now has the following two effects on functions:

- The deprecation notice is shown at runtime when the decorated function is called.
- The deprecation notice is shown on the Sphinx documentation (see examples below).

Authors:
  - Tianyu Liu (https://github.com/kingcrimsontianyu)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)
  - Tom Augspurger (https://github.com/TomAugspurger)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/kvikio/pull/714
