+++ python/kvikio/kvikio/utils.py
@@ -131,8 +131,11 @@ def call_once(func: Callable) -> Callable:
     return wrapper
 
 
-def kvikio_deprecation_notice(msg: str) -> Callable:
-    """Decorate a function to print the deprecation notice at runtime.
+def kvikio_deprecation_notice(
+    msg: str = "This function is deprecated.", *, since: str
+) -> Callable:
+    """Decorate a function to print the deprecation notice at runtime,
+       and also add the notice to Sphinx documentation.
 
     Examples:
 
@@ -146,6 +149,9 @@ def kvikio_deprecation_notice(msg: str) -> Callable:
     msg: str
         The deprecation notice.
 
+    since: str
+        The KvikIO version since which the function becomes deprecated.
+
     Returns
     -------
     Callable
@@ -157,8 +163,28 @@ def kvikio_deprecation_notice(msg: str) -> Callable:
             warnings.warn(msg, category=FutureWarning, stacklevel=2)
             return func(*args, **kwargs)
 
-        # Allow the docstring to be corrected generated for the decorated func in Sphinx
-        wrapper.__doc__ = func.__doc__
+        # Allow the docstring to be correctly generated for the decorated func in Sphinx
+        func_doc = getattr(func, "__doc__")
+        valid_docstring = "" if func_doc is None else func_doc
+        wrapper.__doc__ = "{:} {:} {:}\n\n    {:}".format(
+            ".. deprecated::", since, msg, valid_docstring
+        )
+
         return wrapper
 
     return decorator
+
+
+def kvikio_deprecate_module(msg: str = "", *, since: str) -> None:
+    """Mark a module as deprecated.
+
+    Parameters
+    ----------
+    msg: str
+        The deprecation notice.
+
+    since: str
+        The KvikIO version since which the module becomes deprecated.
+    """
+    full_msg = f"This module is deprecated since {since}. {msg}"
+    warnings.warn(full_msg, category=FutureWarning, stacklevel=2)
